Fix wrong path element filtering
Previously I only filtered based on the fact if a line includes "d="
which also matched elements that only had id, not id. To ensure that
only path elements including "d=" (not <path />) elements will be
left after filtering I also started to check if the element was path.
(Not only if it has " d=" property becuase glyph and other elements
have "d" attribute as well.)