Finish game loop
If you invide a country and you answer is correct, you conquere the
country, otherwise one of the neighboring countries of the invaded
country will be removed from conquered (the country neighbor country
that you invided first will be the one removed). If you have none
conquered countries left, you lose and the game restarts. If you
conquered all the countrie, you win and the game restarts.